+++ dotnet/OLI.NetCli/Program.cs
@@ -804,709 +804,7 @@ class Program
         }, importTasksPathOpt);
 
 
-        var memoryInfoCmd = new Command("memory-info", "Show memory file path and content");
-        memoryInfoCmd.SetHandler(async () =>
-        {
-            if (File.Exists(MemoryPath))
-            {
-                Console.WriteLine($"Memory path: {MemoryPath}");
-                Console.WriteLine(File.ReadAllText(MemoryPath));
-            }
-            else
-            {
-                Console.WriteLine("No memory file found");
-            }
-            await Task.CompletedTask;
-        });
 
-        var sectionOption = new Option<string>("--section") { IsRequired = true };
-        var memoryOption = new Option<string>("--memory") { IsRequired = true };
-        var addMemoryCmd = new Command("add-memory", "Add a memory entry")
-        {
-            sectionOption,
-            memoryOption
-        };
-        addMemoryCmd.SetHandler(async (string section, string memory) =>
-        {
-            var content = File.Exists(MemoryPath) ? File.ReadAllText(MemoryPath) : "";
-            var sectionHeader = $"## {section}";
-            if (!content.Contains(sectionHeader))
-            {
-                content += $"\n{sectionHeader}\n";
-            }
-            var insertIndex = content.IndexOf(sectionHeader) + sectionHeader.Length;
-            content = content.Insert(insertIndex, $"\n- {memory}");
-            File.WriteAllText(MemoryPath, content);
-            Console.WriteLine("Memory added");
-            await Task.CompletedTask;
-        }, sectionOption, memoryOption);
-
-        var contentOption = new Option<string>("--content") { IsRequired = true };
-        var replaceMemoryCmd = new Command("replace-memory-file", "Replace entire memory file")
-        {
-            contentOption
-        };
-        replaceMemoryCmd.SetHandler(async (string content) =>
-        {
-            File.WriteAllText(MemoryPath, content);
-            Console.WriteLine("Memory file replaced");
-            await Task.CompletedTask;
-        }, contentOption);
-
-        var appendContentOpt = new Option<string>("--content") { IsRequired = true };
-        var appendMemoryCmd = new Command("append-memory-file", "Append text to memory file")
-        {
-            appendContentOpt
-        };
-        appendMemoryCmd.SetHandler(async (string content) =>
-        {
-            var existing = File.Exists(MemoryPath) ? File.ReadAllText(MemoryPath) : string.Empty;
-            File.WriteAllText(MemoryPath, existing + content);
-            Console.WriteLine("Memory file updated");
-            await Task.CompletedTask;
-        }, appendContentOpt);
-
-        var importMemoryPathOpt = new Option<string>("--path") { IsRequired = true };
-        var importMemoryCmd = new Command("import-memory-file", "Load memory file from path")
-        {
-            importMemoryPathOpt
-        };
-        importMemoryCmd.SetHandler(async (string path) =>
-        {
-            if (!File.Exists(path))
-            {
-                Console.WriteLine("File not found");
-                return;
-            }
-            File.Copy(path, MemoryPath, true);
-            Console.WriteLine("Memory file imported");
-            await Task.CompletedTask;
-        }, importMemoryPathOpt);
-
-        var exportMemoryPathOpt = new Option<string>("--path") { IsRequired = true };
-        var exportMemoryCmd = new Command("export-memory-file", "Write memory file to path")
-        {
-            exportMemoryPathOpt
-        };
-        exportMemoryCmd.SetHandler(async (string path) =>
-        {
-            if (!File.Exists(MemoryPath))
-            {
-                Console.WriteLine("No memory file found");
-                return;
-            }
-            File.Copy(MemoryPath, path, true);
-            Console.WriteLine($"Memory file exported to {path}");
-            await Task.CompletedTask;
-        }, exportMemoryPathOpt);
-
-        var linesStartOpt = new Option<int>("--start", () => 1);
-        var linesEndOpt = new Option<int>("--end", () => int.MaxValue);
-        var memoryLinesCmd = new Command("memory-lines", "Show lines from memory file") { linesStartOpt, linesEndOpt };
-        memoryLinesCmd.SetHandler(async (int start, int end) =>
-        {
-            if (!File.Exists(MemoryPath))
-            {
-                Console.WriteLine("No memory file found");
-                return;
-            }
-            var lines = File.ReadAllLines(MemoryPath);
-            start = Math.Max(1, start);
-            end = Math.Min(lines.Length, end);
-            for (int i = start; i <= end; i++)
-            {
-                Console.WriteLine($"{i}: {lines[i - 1]}");
-            }
-            await Task.CompletedTask;
-        }, linesStartOpt, linesEndOpt);
-
-        var memHeadLinesOpt = new Option<int>("--lines", () => 10);
-        var memoryHeadCmd = new Command("memory-head", "Show first lines of memory") { memHeadLinesOpt };
-        memoryHeadCmd.SetHandler(async (int lines) =>
-        {
-            Console.WriteLine(MemoryUtils.Head(MemoryPath, lines));
-            await Task.CompletedTask;
-        }, memHeadLinesOpt);
-
-        var memTailLinesOpt = new Option<int>("--lines", () => 10);
-        var memoryTailCmd = new Command("memory-tail", "Show last lines of memory") { memTailLinesOpt };
-        memoryTailCmd.SetHandler(async (int lines) =>
-        {
-            Console.WriteLine(MemoryUtils.Tail(MemoryPath, lines));
-            await Task.CompletedTask;
-        }, memTailLinesOpt);
-
-        var insertIndexOpt = new Option<int>("--index") { IsRequired = true };
-        var insertTextOpt = new Option<string>("--text") { IsRequired = true };
-        var insertMemoryCmd = new Command("insert-memory-lines", "Insert lines into memory file") { insertIndexOpt, insertTextOpt };
-        insertMemoryCmd.SetHandler(async (int index, string text) =>
-        {
-            var lines = File.Exists(MemoryPath) ? File.ReadAllLines(MemoryPath).ToList() : new List<string>();
-            var newLines = text.Split('\n');
-            index = Math.Clamp(index - 1, 0, lines.Count);
-            lines.InsertRange(index, newLines);
-            File.WriteAllLines(MemoryPath, lines);
-            Console.WriteLine("Memory updated");
-            await Task.CompletedTask;
-        }, insertIndexOpt, insertTextOpt);
-
-        var replaceStartOpt = new Option<int>("--start") { IsRequired = true };
-        var replaceEndOpt = new Option<int>("--end") { IsRequired = true };
-        var replaceTextOpt = new Option<string>("--text") { IsRequired = true };
-        var replaceMemoryLinesCmd = new Command("replace-memory-lines", "Replace range of memory lines") { replaceStartOpt, replaceEndOpt, replaceTextOpt };
-        replaceMemoryLinesCmd.SetHandler(async (int start, int end, string text) =>
-        {
-            if (!File.Exists(MemoryPath))
-            {
-                Console.WriteLine("No memory file found");
-                return;
-            }
-            var lines = File.ReadAllLines(MemoryPath).ToList();
-            start = Math.Max(1, start);
-            end = Math.Min(lines.Count, end);
-            lines.RemoveRange(start - 1, Math.Max(0, end - start + 1));
-            lines.InsertRange(start - 1, text.Split('\n'));
-            File.WriteAllLines(MemoryPath, lines);
-            Console.WriteLine("Memory updated");
-            await Task.CompletedTask;
-        }, replaceStartOpt, replaceEndOpt, replaceTextOpt);
-
-        var mergeMemoryPathOpt = new Option<string>("--path") { IsRequired = true };
-        var mergeMemoryCmd = new Command("merge-memory-file", "Merge another memory file")
-        {
-            mergeMemoryPathOpt
-        };
-        mergeMemoryCmd.SetHandler(async (string path) =>
-        {
-            if (!File.Exists(path))
-            {
-                Console.WriteLine("File not found");
-                return;
-            }
-            var existing = File.Exists(MemoryPath) ? File.ReadAllText(MemoryPath) : string.Empty;
-            var extra = File.ReadAllText(path);
-            File.WriteAllText(MemoryPath, existing + extra);
-            Console.WriteLine("Memory files merged");
-            await Task.CompletedTask;
-        }, mergeMemoryPathOpt);
-
-        var resetMemoryCmd = new Command("reset-memory-file", "Delete and recreate memory file");
-        resetMemoryCmd.SetHandler(async () =>
-        {
-            var template = "# oli.md\n\n## Project Structure\n";
-            File.WriteAllText(MemoryPath, template);
-            Console.WriteLine("Memory file reset");
-            await Task.CompletedTask;
-        });
-
-        var statePathCmd = new Command("state-path", "Show path of state file");
-        statePathCmd.SetHandler(async () =>
-        {
-            Console.WriteLine(StatePath);
-            await Task.CompletedTask;
-        });
-
-        var stateInfoCmd = new Command("state-info", "Show summary of saved state");
-        stateInfoCmd.SetHandler(async () =>
-        {
-            var state = LoadState();
-            Console.WriteLine($"AgentMode:{state.AgentMode} Model:{state.SelectedModel} Tasks:{state.Tasks.Count} Messages:{state.Conversation.Count}");
-            await Task.CompletedTask;
-        });
-
-        var stateVersionCmd = new Command("state-version", "Show state file version");
-        stateVersionCmd.SetHandler(async () =>
-        {
-            var state = LoadState();
-            Console.WriteLine(state.StateVersion);
-            await Task.CompletedTask;
-        });
-
-        var stateSummaryCmd = new Command("state-summary", "Show counts of tasks, messages, and summaries");
-        stateSummaryCmd.SetHandler(async () =>
-        {
-            var state = LoadState();
-            Console.WriteLine($"tasks:{state.Tasks.Count} messages:{state.Conversation.Count} summaries:{state.ConversationSummaries.Count}");
-            await Task.CompletedTask;
-        });
-
-        var stateFilesCmd = new Command("state-files", "List all state file paths");
-        stateFilesCmd.SetHandler(async () =>
-        {
-            Console.WriteLine(StatePath);
-            Console.WriteLine(TasksPath);
-            Console.WriteLine(ConversationPath);
-            Console.WriteLine(SummariesPath);
-            Console.WriteLine(ToolsPath);
-            Console.WriteLine(LspPath);
-            await Task.CompletedTask;
-        });
-
-        var setWorkingDirOpt = new Option<string>("--path") { IsRequired = true };
-        var setWorkingDirCmd = new Command("set-working-dir", "Set working directory") { setWorkingDirOpt };
-        setWorkingDirCmd.SetHandler(async (string path) =>
-        {
-            var state = LoadState();
-            state.WorkingDirectory = path;
-            SaveState(state);
-            Console.WriteLine($"Working directory set to {path}");
-            await Task.CompletedTask;
-        }, setWorkingDirOpt);
-
-        var currentDirCmd = new Command("current-directory", "Show configured working directory");
-        currentDirCmd.SetHandler(async () =>
-        {
-            var state = LoadState();
-            Console.WriteLine(state.WorkingDirectory);
-            await Task.CompletedTask;
-        });
-
-        var memoryPathCmd = new Command("memory-path", "Show path of memory file");
-        memoryPathCmd.SetHandler(async () =>
-        {
-            Console.WriteLine(MemoryPath);
-            await Task.CompletedTask;
-        });
-
-        var memoryExistsCmd = new Command("memory-exists", "Check for memory file");
-        memoryExistsCmd.SetHandler(async () =>
-        {
-            Console.WriteLine(File.Exists(MemoryPath) ? "true" : "false");
-            await Task.CompletedTask;
-        });
-
-        var createMemoryCmd = new Command("create-memory-file", "Create memory file if missing");
-        createMemoryCmd.SetHandler(async () =>
-        {
-            if (!File.Exists(MemoryPath))
-            {
-                var template = "# oli.md\n\n## Project Structure\n";
-                File.WriteAllText(MemoryPath, template);
-                Console.WriteLine("Memory file created");
-            }
-            else
-            {
-                Console.WriteLine("Memory file already exists");
-            }
-            await Task.CompletedTask;
-        });
-
-        var parseMemoryCmd = new Command("parsed-memory", "Show parsed memory sections");
-        parseMemoryCmd.SetHandler(async () =>
-        {
-            if (!File.Exists(MemoryPath))
-            {
-                Console.WriteLine("No memory file found");
-                return;
-            }
-            var lines = File.ReadAllLines(MemoryPath);
-            string? current = null;
-            var map = new Dictionary<string, List<string>>();
-            foreach (var line in lines)
-            {
-                if (line.StartsWith("## "))
-                {
-                    current = line[3..];
-                    map[current] = new List<string>();
-                }
-                else if (line.StartsWith("- ") && current != null)
-                {
-                    map[current].Add(line[2..]);
-                }
-            }
-            foreach (var kv in map)
-            {
-                Console.WriteLine($"Section: {kv.Key}");
-                foreach (var entry in kv.Value)
-                {
-                    Console.WriteLine($"  - {entry}");
-                }
-            }
-            await Task.CompletedTask;
-        });
-
-        var copySectionOpt = new Option<string>("--section") { IsRequired = true };
-        var copyDestOpt = new Option<string>("--dest") { IsRequired = true };
-        var copySectionCmd = new Command("copy-memory-section", "Copy memory section to file") { copySectionOpt, copyDestOpt };
-        copySectionCmd.SetHandler(async (string section, string dest) =>
-        {
-            if (!File.Exists(MemoryPath)) { Console.WriteLine("No memory file"); return; }
-            var lines = File.ReadAllLines(MemoryPath);
-            var collecting = false;
-            var selected = new List<string>();
-            foreach (var line in lines)
-            {
-                if (line.StartsWith("## "))
-                {
-                    collecting = line[3..].Trim() == section;
-                    continue;
-                }
-                if (collecting && line.StartsWith("- "))
-                {
-                    selected.Add(line);
-                }
-            }
-            File.WriteAllLines(dest, selected);
-            Console.WriteLine($"Section copied to {dest}");
-            await Task.CompletedTask;
-        }, copySectionOpt, copyDestOpt);
-
-        var swapAOpt = new Option<string>("--first") { IsRequired = true };
-        var swapBOpt = new Option<string>("--second") { IsRequired = true };
-        var swapSectionCmd = new Command("swap-memory-sections", "Swap two memory sections") { swapAOpt, swapBOpt };
-        swapSectionCmd.SetHandler(async (string first, string second) =>
-        {
-            if (!File.Exists(MemoryPath)) { Console.WriteLine("No memory file"); return; }
-            var lines = File.ReadAllLines(MemoryPath).ToList();
-            int idxA = lines.FindIndex(l => l.StartsWith("## "+first));
-            int idxB = lines.FindIndex(l => l.StartsWith("## "+second));
-            if (idxA == -1 || idxB == -1) { Console.WriteLine("Section not found"); return; }
-            if (idxA > idxB) { var t = idxA; idxA = idxB; idxB = t; var name = first; first = second; second = name; }
-            int endA = lines.Skip(idxA+1).TakeWhile(l => !l.StartsWith("## ")).Count();
-            int endB = lines.Skip(idxB+1).TakeWhile(l => !l.StartsWith("## ")).Count();
-            var secA = lines.GetRange(idxA, endA+1);
-            var secB = lines.GetRange(idxB, endB+1);
-            lines.RemoveRange(idxB, endB+1);
-            lines.InsertRange(idxB, secA);
-            lines.RemoveRange(idxA, endA+1);
-            lines.InsertRange(idxA, secB);
-            File.WriteAllLines(MemoryPath, lines);
-            Console.WriteLine("Sections swapped");
-            await Task.CompletedTask;
-        }, swapAOpt, swapBOpt);
-
-        var sectionCountCmd = new Command("memory-section-count", "Count memory sections");
-        sectionCountCmd.SetHandler(async () =>
-        {
-            if (!File.Exists(MemoryPath))
-            {
-                Console.WriteLine("0");
-                return;
-            }
-            var count = File.ReadLines(MemoryPath).Count(l => l.StartsWith("## "));
-            Console.WriteLine(count);
-            await Task.CompletedTask;
-        });
-
-        var entryCountCmd = new Command("memory-entry-count", "Count total memory entries");
-        entryCountCmd.SetHandler(async () =>
-        {
-            if (!File.Exists(MemoryPath))
-            {
-                Console.WriteLine("0");
-                return;
-            }
-            var count = File.ReadLines(MemoryPath).Count(l => l.StartsWith("- "));
-            Console.WriteLine(count);
-            await Task.CompletedTask;
-        });
-
-        var memoryTemplateCmd = new Command("memory-template", "Show default memory template");
-        memoryTemplateCmd.SetHandler(async () =>
-        {
-            Console.WriteLine("# oli.md\n\n## Project Structure\n- example\n\n## Build Commands\n- example\n\n## Test Commands\n- example\n\n## Architecture\n- example");
-            await Task.CompletedTask;
-        });
-
-        var memorySizeCmd = new Command("memory-size", "Show memory file size");
-        memorySizeCmd.SetHandler(async () =>
-        {
-            var size = File.Exists(MemoryPath) ? new FileInfo(MemoryPath).Length : 0;
-            Console.WriteLine(size);
-            await Task.CompletedTask;
-        });
-
-        var memSearchTextOpt = new Option<string>("--text") { IsRequired = true };
-        var searchMemoryCmd = new Command("search-memory", "Search memory for text") { memSearchTextOpt };
-        searchMemoryCmd.SetHandler(async (string text) =>
-        {
-            if (!File.Exists(MemoryPath))
-            {
-                Console.WriteLine("No memory file found");
-                return;
-            }
-            foreach (var line in File.ReadLines(MemoryPath))
-            {
-                if (line.Contains(text, StringComparison.OrdinalIgnoreCase))
-                {
-                    Console.WriteLine(line);
-                }
-            }
-            await Task.CompletedTask;
-        }, memSearchTextOpt);
-
-        var deletePatternOpt = new Option<string>("--pattern") { IsRequired = true };
-        var deleteMemoryLineCmd = new Command("delete-memory-lines", "Remove memory lines matching pattern") { deletePatternOpt };
-        deleteMemoryLineCmd.SetHandler(async (string pattern) =>
-        {
-            if (!File.Exists(MemoryPath))
-            {
-                Console.WriteLine("No memory file found");
-                return;
-            }
-            var lines = File.ReadAllLines(MemoryPath).ToList();
-            var remaining = lines.Where(l => !l.Contains(pattern, StringComparison.OrdinalIgnoreCase)).ToList();
-            File.WriteAllLines(MemoryPath, remaining);
-            Console.WriteLine(lines.Count - remaining.Count);
-            await Task.CompletedTask;
-        }, deletePatternOpt);
-
-
-
-        var currentModelCmd = new Command("current-model", "Show selected model index");
-        currentModelCmd.SetHandler(() =>
-        {
-            var state = LoadState();
-            Console.WriteLine(state.SelectedModel);
-            return Task.CompletedTask;
-        });
-
-        var listSubsCmd = new Command("subscriptions", "List active event subscriptions");
-        listSubsCmd.SetHandler(() =>
-        {
-            var state = LoadState();
-            if (state.Subscriptions.Count == 0)
-            {
-                Console.WriteLine("No subscriptions");
-            }
-            foreach (var s in state.Subscriptions)
-            {
-                Console.WriteLine(s);
-            }
-            return Task.CompletedTask;
-        });
-
-        var subscriptionCountCmd = new Command("subscription-count", "Number of active subscriptions");
-        subscriptionCountCmd.SetHandler(() =>
-        {
-            var state = LoadState();
-            Console.WriteLine(state.Subscriptions.Count);
-            return Task.CompletedTask;
-        });
-
-        var deleteSectionOption = new Option<string>("--section") { IsRequired = true };
-        var deleteMemorySectionCmd = new Command("delete-memory-section", "Remove a memory section")
-        {
-            deleteSectionOption
-        };
-        deleteMemorySectionCmd.SetHandler(async (string section) =>
-        {
-            if (!File.Exists(MemoryPath))
-            {
-                Console.WriteLine("No memory file found");
-                return;
-            }
-            var lines = File.ReadAllLines(MemoryPath);
-            var result = new List<string>();
-            bool skip = false;
-            foreach (var line in lines)
-            {
-                if (line.StartsWith("## "))
-                {
-                    skip = line[3..] == section;
-                    if (!skip) result.Add(line);
-                    continue;
-                }
-                if (!skip) result.Add(line);
-            }
-            File.WriteAllLines(MemoryPath, result);
-            Console.WriteLine($"Deleted section {section}");
-            await Task.CompletedTask;
-        }, deleteSectionOption);
-
-        var deleteIdOption = new Option<string>("--id") { IsRequired = true };
-        var deleteTaskCmd = new Command("delete-task", "Remove a task")
-        {
-            deleteIdOption
-        };
-        deleteTaskCmd.SetHandler(async (string id) =>
-        {
-            var state = LoadState();
-            var idx = state.Tasks.FindIndex(t => t.Id == id);
-            if (idx >= 0)
-            {
-                state.Tasks.RemoveAt(idx);
-                SaveState(state);
-                Console.WriteLine("Task removed");
-            }
-            else
-            {
-                Console.WriteLine("Task not found");
-            }
-            await Task.CompletedTask;
-        }, deleteIdOption);
-
-        var infoIdOption = new Option<string>("--id") { IsRequired = true };
-        var taskInfoCmd = new Command("task-info", "Show details for a task")
-        {
-            infoIdOption
-        };
-        taskInfoCmd.SetHandler(async (string id) =>
-        {
-            var state = LoadState();
-            var task = state.Tasks.Find(t => t.Id == id);
-            if (task != null)
-            {
-                Console.WriteLine($"Id: {task.Id}\nDescription: {task.Description}\nStatus: {task.Status}");
-                if (task.DueDate != null) Console.WriteLine($"Due: {task.DueDate:u}");
-                if (task.Tags.Count > 0) Console.WriteLine($"Tags: {string.Join(',', task.Tags)}");
-                Console.WriteLine($"Tools: {task.ToolCount}\nInput Tokens: {task.InputTokens}\nOutput Tokens: {task.OutputTokens}");
-            }
-            else
-            {
-                Console.WriteLine("Task not found");
-            }
-            await Task.CompletedTask;
-        }, infoIdOption);
-
-        var latestTaskCmd = new Command("latest-task", "Show most recently created task");
-        latestTaskCmd.SetHandler(async () =>
-        {
-            var state = LoadState();
-            var task = state.Tasks.OrderByDescending(t => t.CreatedAt).FirstOrDefault();
-            if (task != null)
-            {
-                Console.WriteLine($"{task.Id}: {task.Description} [{task.Status}]");
-            }
-            else
-            {
-                Console.WriteLine("No tasks");
-            }
-            await Task.CompletedTask;
-        });
-
-        var inProgressCmd = new Command("tasks-in-progress", "List IDs of in-progress tasks");
-        inProgressCmd.SetHandler(async () =>
-        {
-            var state = LoadState();
-            foreach (var t in state.Tasks.Where(t => t.Status == "in-progress"))
-                Console.WriteLine(t.Id);
-            await Task.CompletedTask;
-        });
-
-        var taskDescCmd = new Command("task-descriptions", "List task descriptions");
-        taskDescCmd.SetHandler(async () =>
-        {
-            var state = LoadState();
-            foreach (var t in state.Tasks)
-                Console.WriteLine(t.Description);
-            await Task.CompletedTask;
-        });
-
-        var resetStateCmd = new Command("reset-state", "Delete saved state file");
-        resetStateCmd.SetHandler(async () =>
-        {
-            if (File.Exists(StatePath))
-            {
-                File.Delete(StatePath);
-                Console.WriteLine("State reset");
-            }
-            else
-            {
-                Console.WriteLine("No state file");
-            }
-            await Task.CompletedTask;
-        });
-
-        var importPathOption = new Option<string>("--path") { IsRequired = true };
-        var importStateCmd = new Command("import-state", "Load state from file")
-        {
-            importPathOption
-        };
-        importStateCmd.SetHandler(async (string path) =>
-        {
-            if (!File.Exists(path))
-            {
-                Console.WriteLine("File not found");
-                return;
-            }
-            var json = File.ReadAllText(path);
-            var state = JsonSerializer.Deserialize<AppState>(json);
-            if (state != null)
-            {
-                SaveState(state);
-                Console.WriteLine("State imported");
-            }
-            else
-            {
-                Console.WriteLine("Invalid state file");
-            }
-            await Task.CompletedTask;
-        }, importPathOption);
-
-        var exportPathOption = new Option<string>("--path") { IsRequired = true };
-        var exportStateCmd = new Command("export-state", "Write state to file")
-        {
-            exportPathOption
-        };
-        exportStateCmd.SetHandler(async (string path) =>
-        {
-            var state = LoadState();
-            File.WriteAllText(path, JsonSerializer.Serialize(state, new JsonSerializerOptions { WriteIndented = true }));
-            Console.WriteLine($"State exported to {path}");
-            await Task.CompletedTask;
-        }, exportPathOption);
-
-        var deleteMemoryFileCmd = new Command("delete-memory-file", "Remove memory file");
-        deleteMemoryFileCmd.SetHandler(async () =>
-        {
-            if (File.Exists(MemoryPath))
-            {
-                File.Delete(MemoryPath);
-                Console.WriteLine("Memory file deleted");
-            }
-            else
-            {
-                Console.WriteLine("No memory file found");
-            }
-            await Task.CompletedTask;
-        });
-
-        var listMemorySectionsCmd = new Command("list-memory-sections", "List memory sections");
-        listMemorySectionsCmd.SetHandler(async () =>
-        {
-            if (!File.Exists(MemoryPath))
-            {
-                Console.WriteLine("No memory file found");
-                return;
-            }
-            foreach (var line in File.ReadLines(MemoryPath))
-            {
-                if (line.StartsWith("## "))
-                {
-                    Console.WriteLine(line[3..]);
-                }
-            }
-            await Task.CompletedTask;
-        });
-
-        var tasksPathCmd = new Command("tasks-path", "Show tasks file path");
-        tasksPathCmd.SetHandler(async () =>
-        {
-            Console.WriteLine(TasksPath);
-            await Task.CompletedTask;
-        });
-
-        var conversationPathCmd = new Command("conversation-path", "Show conversation file path");
-        conversationPathCmd.SetHandler(async () =>
-        {
-            Console.WriteLine(ConversationPath);
-            await Task.CompletedTask;
-        });
-
-        var summariesPathCmd = new Command("summaries-path", "Show summaries file path");
-        summariesPathCmd.SetHandler(async () =>
-        {
-            Console.WriteLine(SummariesPath);
-            await Task.CompletedTask;
-        });
-
-        var toolsPathCmd = new Command("tools-path", "Show tools file path");
-        toolsPathCmd.SetHandler(async () =>
-        {
-            Console.WriteLine(ToolsPath);
-            await Task.CompletedTask;
-        });
 
         var versionCmd = new Command("version", "Display CLI version");
         versionCmd.SetHandler(async () =>
@@ -1558,31 +856,29 @@ class Program
 
         var root = new RootCommand("oli .NET CLI")
         {
-            runCmd, agentCmd, agentStatusCmd, setModelCmd, modelsCmd,
-            memoryInfoCmd, memoryPathCmd, createMemoryCmd,
-            addMemoryCmd, replaceMemoryCmd, parseMemoryCmd,
-            sectionCountCmd, entryCountCmd, memoryTemplateCmd, memorySizeCmd, searchMemoryCmd, deleteMemoryLineCmd, mergeMemoryCmd, resetMemoryCmd, copySectionCmd, swapSectionCmd, memoryLinesCmd, memoryHeadCmd, memoryTailCmd, insertMemoryCmd, replaceMemoryLinesCmd,
-            setAutoCompressCmd, setThresholdsCmd,
-            readFileCmd, readNumberedCmd, readLinesCmd,
-            writeFileCmd, writeDiffCmd, editFileCmd, appendFileCmd,
-            genWriteDiffCmd, genEditDiffCmd, copyFileCmd, moveFileCmd, renameFileCmd,
-            deleteFileCmd, fileExistsCmd, touchFileCmd, listDirCmd, listDirRecursiveCmd, headFileCmd, tailFileCmd, fileSizeCmd, createDirCmd, deleteDirCmd, dirExistsCmd, copyDirCmd, moveDirCmd, renameDirCmd, fileInfoCmd, countLinesCmd,
-            globSearchCmd, globSearchInDirCmd, grepSearchCmd,
-            currentModelCmd, listSubsCmd, deleteMemorySectionCmd,
-            resetStateCmd,
-            importStateCmd, exportStateCmd, deleteMemoryFileCmd,
-            listMemorySectionsCmd, tasksPathCmd, conversationPathCmd, summariesPathCmd, toolsPathCmd,
-            appendMemoryCmd, importMemoryCmd, exportMemoryCmd, statePathCmd, stateInfoCmd, stateVersionCmd, stateSummaryCmd, stateFilesCmd, versionCmd,
-            memoryExistsCmd, subscribeCmd, unsubscribeCmd, subscriptionCountCmd,
+            runCmd,
+            agentCmd,
+            agentStatusCmd,
+            setModelCmd,
+            modelsCmd,
+            versionCmd,
+            subscribeCmd,
+            unsubscribeCmd,
+            subscriptionCountCmd,
             deleteSummaryRangeCmd,
-            setWorkingDirCmd, currentDirCmd,
-            runCommandCmd, rpcStartCmd, rpcStopCmd, rpcStatusCmd, rpcNotifyCmd
+            runCommandCmd,
+            rpcStartCmd,
+            rpcStopCmd,
+            rpcStatusCmd,
+            rpcNotifyCmd
         };
 
         LspCommands.Register(root);
         TaskCommands.Register(root);
         ToolCommands.Register(root);
         ConversationCommands.Register(root);
+        MemoryCommands.Register(root);
+        StateCommands.Register(root);
         AdditionalCommands.Register(root);
         FileCommands.Register(root);
         SummaryCommands.Register(root);
